requirements.txt
# Core dependencies
streamlit==1.29.0
pandas
plotly==5.18.0
python-dotenv==1.0.0

# API and web
fastapi==0.108.0
uvicorn==0.25.0
anthropic==0.18.1

# Data processing
numpy==1.26.2
scipy

# Geographic visualization
folium==0.15.1
streamlit-folium==0.15.1

# Additional utilities
requests==2.31.0
pydantic